requirements.txt
# Core dependencies
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
implicit>=0.7.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Database
# sqlite3

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
